Nick-Assistant
 Updated 24 seconds ago
So when i was learning about python from various tutorials at that time i made this project .It is basically (NLP) AI assistant which do some work like send mail and various as ,I Mention bellow so in this. I used (Pyaudio modual ) in most of area it is totally designed and wrote in python.

What can this A.I. assistant do for you? It can send emails on your behalf. It can play music for you. It can do Wikipedia searches for you. It is capable of opening websites like Google, Youtube, etc., in a web browser. It is capable of opening your code editor or IDE with a single voice command.